+++ semantix/__init__.py
@@ -81,5 +81,5 @@ try:
 except ImportError:
     pass
 
-__version__ = "0.3.0"
+__version__ = "0.3.1"
 __author__ = "Akhona Eland"